requirements.txt
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
